# openapi.yaml
openapi: "3.0.2"
info:
  description: A farm for abandoned UUIDs
  title: uuid farm
  version: v1
servers:
  - url: http://uuid-farm.io/v1
    description: uuid farm
paths:
  /:
    get:
      summary: Farm stats
      operationId: getFarm
      responses:
        "200":
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  uuids:
                    type: integer
                    example: 10
  /uuids:
    get:
      summary: Get uuids
      description: Get all uuids.
      operationId: getUUIDs
      responses:
        "200":
          description: uuid array response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/uuid"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
    post:
      summary: Adopt uuid
      description: Adopt new uuid that's on the farm.
      operationId: adoptUUID
      parameters:
        - name: id
          in: path
          description: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: uuid response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uuid"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
  /uuids/{id}:
    get:
      summary: Get uuid that's on the farm
      description: Get a uuid that's on the farm.
      operationId: getUUID
      parameters:
        - name: id
          in: path
          description: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: uuid response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uuid"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
    post:
      summary: Surrender uuid
      description: Give uuid up for adoption.
      operationId: surrenderUUID
      parameters:
        - name: id
          in: path
          description: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: uuid response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uuid"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
    put:
      summary: Update uuid
      description: Update uuid that's on the farm.
      operationId: updateUUID
      parameters:
        - name: id
          in: path
          description: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: uuid
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/uuid"
      responses:
        "200":
          description: uuid response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uuid"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
components:
  schemas:
    uuid:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        tag:
          type: string
        surrenderDate:
          type: string
          format: date
        adoptionDate:
          type: string
          format: date
    apiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 400
          format: int32
        message:
          type: string
          example: Bad request.